% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage.R
\name{covrpage}
\alias{covrpage}
\title{render covrpage}
\usage{
covrpage(pkg)
}
\arguments{
\item{pkg}{path to package}
}
\description{
Render the template of the covrpage in the package directory.
}
\seealso{
\code{\link[rmarkdown]{render}}
}
