% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covrpage_snapshot.R
\name{covrpage_snapshot}
\alias{covrpage_snapshot}
\title{Create covrpages for pacakges on Github}
\usage{
covrpage_snapshot(repo, ...)
}
\arguments{
\item{repo}{character, user/repo}

\item{...}{arguments to pass to coverpage}
}
\description{
Preview covrpages for R package repositories on Github
}
\details{
This function only works for public repositories.
}
\examples{
\dontrun{
if(interactive()){
 covrpage_snapshot(repo = 'hrbrmstr/slackr')
 }
}
}
