% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_covrpage.R
\name{travis_encrypt}
\alias{travis_encrypt}
\title{encrypt travis environment variables}
\usage{
travis_encrypt(r_obj = Sys.getenv("GITHUB_PAT"), travis_env = "GH_PAT",
  add = TRUE, test = FALSE)
}
\arguments{
\item{r_obj}{object value to encrypt, Default: Sys.getenv("GITHUB_PAT")}

\item{travis_env}{character, name of travis environment variable, Default: 'GH_PAT'}

\item{add}{boolean, add directly to .travis.yml, Default: TRUE}

\item{test}{boolean, print example of command to be run to console, Default: FALSE}
}
\description{
add encrypted travis environment variable to git directory and .travis.yml
}
